Extract ShakeEffect and ElapsedTimeTracker into own files
