+++ CodeWordBreaker/CodeWordBreaker/UI/Extensions/ShakeEffect.swift
@@ -1,0 +1,29 @@
+//
+//  ShakeEffect.swift
+//  CodeWordBreaker
+//
+//  Created by Alexander Ostrovsky on 12/4/2026.
+//
+
+import SwiftUI
+
+struct ShakeEffect: GeometryEffect {
+    private static let cyclesPerShake: CGFloat = 3
+    private static let amplitude: CGFloat = 10
+
+    private var shakes: CGFloat
+
+    init(shakes: Int) {
+        self.shakes = CGFloat(shakes)
+    }
+
+    var animatableData: CGFloat {
+        get { shakes }
+        set { shakes = newValue }
+    }
+
+    func effectValue(size: CGSize) -> ProjectionTransform {
+        let translation = sin(shakes * .pi * 2 * Self.cyclesPerShake) * Self.amplitude
+        return ProjectionTransform(CGAffineTransform(translationX: translation, y: 0))
+    }
+}
